refactor: Simplify the directory structure
