+++ .vscode/settings.json
@@ -13,13 +13,13 @@
     "grep": true,
     "head": true,
     "ls": true,
-    "tail": true,
-    "pnpm run check": true,
-    "pnpm run fix": true,
     "pnpm check": true,
     "pnpm fix": true,
+    "pnpm run check": true,
+    "pnpm run fix": true,
+    "pnpm test": true,
     "pnpm vitest": true,
-    "pnpm test": true
+    "tail": true
   },
   "editor.codeActionsOnSave": {
     "source.addMissingImports.ts": "always",
@@ -29,6 +29,5 @@
     "source.organizeLinkDefinitions": "always",
     "source.sort.json": "always",
     "source.sortImports": "always"
-  },
-  "typescript.tsdk": "node_modules/typescript/lib"
+  }
 }
