+++ src/frontend/SerialManagerTransport.ts
