+++ src/frontend/components/PortDisconnectedBanner.tsx
