+++ src/frontend/components/ReadPanel.tsx
