+++ src/frontend/index.html
@@ -5,12 +5,12 @@
   <meta charset="UTF-8">
   <meta name="viewport" content="width=device-width, initial-scale=1.0">
   <title>Modbus Web Monitor</title>
-  <link rel="stylesheet" href="./src/style.css">
+  <link rel="stylesheet" href="./style.css">
 </head>
 
 <body>
   <div id="app"></div>
-  <script type="module" src="./src/main.tsx"></script>
+  <script type="module" src="./main.tsx"></script>
 </body>
 
 </html>
