+++ test/components.test.tsx
@@ -2,14 +2,14 @@
 
 import { fireEvent, render, screen } from "@testing-library/preact";
 import { describe, expect, it, vi } from "vitest";
-import { ConnectionSettingsPanel } from "../src/components/ConnectionSettingsPanel.tsx";
-import { DataDisplayPanel } from "../src/components/DataDisplayPanel.tsx";
-import { PortDisconnectedBanner } from "../src/components/PortDisconnectedBanner.tsx";
-import { ReadPanel } from "../src/components/ReadPanel.tsx";
-import { SerialManagerTransport } from "../src/components/SerialManagerTransport.ts";
-import { WritePanel } from "../src/components/WritePanel.tsx";
+import { ConnectionSettingsPanel } from "../src/frontend/components/ConnectionSettingsPanel.tsx";
+import { DataDisplayPanel } from "../src/frontend/components/DataDisplayPanel.tsx";
+import { PortDisconnectedBanner } from "../src/frontend/components/PortDisconnectedBanner.tsx";
+import { ReadPanel } from "../src/frontend/components/ReadPanel.tsx";
+import { WritePanel } from "../src/frontend/components/WritePanel.tsx";
+import { SerialManagerTransport } from "../src/frontend/SerialManagerTransport.ts";
+import type { ModbusResponse } from "../src/modbus.ts";
 import { SerialManager } from "../src/serial.ts";
-import type { ModbusResponse } from "../src/types/modbus.ts";
 
 // Mocks
 class DummySerialManager extends SerialManager {
